게시글 api 수정
datatype 변경

+++ src/main/java/com/vanilla/vanillasns/controlloer/PostController.java
@@ -47,7 +47,7 @@ public class PostController {
 
     //게시글 수정
     @PutMapping("/{id}")
-    public ResponseEntity<Post> updatePost(@PathVariable int id, @RequestBody PostDTO requestDto) {
+    public ResponseEntity<Post> updatePost(@PathVariable long id, @RequestBody PostDTO requestDto) {
         Post updatedPost = postService.updatePost(id, requestDto);
         return new ResponseEntity<>(updatedPost, HttpStatus.OK);
     }
